+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Issue 760](https://github.com/aza547/wow-recorder/issues/760) - Guard against invalid monitor indexes.
 - Add the `-movflags +faststart` ffmpeg flags to the final video cut to improve Cloud load speeds.
 - Fix some Korean translations that were accidentally in Chinese.
+- [Issue 759](https://github.com/aza547/wow-recorder/issues/759) - Fix issue where sleeping windows could trigger a "failed to force stop" error.
 
 ## [7.1.0] - 2025-10-26
 ### Added